% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.abundance.R
\name{ag.abundance}
\alias{ag.abundance}
\title{Aggregate abundance samples}
\usage{
ag.abundance(object, ag.var)
}
\arguments{
\item{object}{A tibble produced by `site.simulate()`}

\item{ag.var}{A column variabe in `object` that is used to form the aggregation}
}
\description{
Takes  site level samples generated by `site.simulate` and aggregates
the abundance by the specified `ag.var` argument
}
\author{
Devin S. Johnson
}
