% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.abundance.R
\name{ag.summ}
\alias{ag.summ}
\title{Summarize aggregated abundance samples}
\usage{
ag.summ(object, ci.prob = 0.95)
}
\arguments{
\item{object}{A tibble produced by `ag.abundance()`}
}
\description{
Takes aggregated abundance samples and summarizes them for plotting, etc.
}
\author{
Devin S. Johnson
}
