% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag.trend.R
\name{ag.trend}
\alias{ag.trend}
\title{Calculate log-linear trend estmates from aggregated abundance data}
\usage{
ag.trend(object, time.range = NULL, ci.prob = 0.95)
}
\arguments{
\item{object}{A tibble produced by `ag.abundance()`}

\item{time.range}{A length 2 vector giving the start and end times for the
trend estimates. These times are between 0 and the number times in the data set.
Not, for example, specific years.}
}
\description{
Takes aggregated abundance values calculated from `ag.abundance` and
calculates trend estimates for each region
}
\author{
Devin S. Johnson
}
