% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.agtrend.data.R
\name{make.agtrend.data}
\alias{make.agtrend.data}
\title{Produce data object for use in \code{site.simulate}}
\usage{
make.agtrend.data(object, abundance.name, site.name, time.name, gam.cut)
}
\arguments{
\item{gam.cut}{}
}
\value{
A tibble
}
\description{
This function creates the specific data form used my the remaining functions in the
package for simulating missing information at all sites.
}
\author{
Devin S. Johnson
}
