% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.fit.gam.R
\name{site.fit.gam}
\alias{site.fit.gam}
\title{Fit a gam model to an individual site data set}
\usage{
site.fit.gam(data, obs.formula = NULL, sig.abund = NULL, min.k = 3,
  ln.adj = 0, family = "tw", obs.prior = NULL)
}
\arguments{
\item{data}{data}

\item{obs.formula}{Formula for the observation model}

\item{sig.abund}{Variance estimates for abundance response if it is estimated.}

\item{min.k}{Minimum df for the thin-plate regression spine model fit to the counts}

\item{ln.adj}{Amount to add to zero valued abundances before taking logs}

\item{obs.prior}{List that provides the mean (named 'gamma.mean') and covariance matrix
for the observation model coefficients (named 'gamma.vcov')}
}
\value{
a list...
}
\description{
The purpose of this function is to allow a user to fit the gam model which is used for
imputation to a sigle site's worth of data. This is primarily used for testing purposes and not meant
to be called by the user in general.
}
\author{
Devin S. Johnson
}
