% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.sim.gam.R
\name{site.sim.gam}
\alias{site.sim.gam}
\title{Simulate abundance measurements from a fitted GAM model}
\usage{
site.sim.gam(fits, data, max.time, size)
}
\arguments{
\item{fits}{Fitted gam object from mgcv}

\item{data}{data}

\item{max.time}{The maximum time used for simulation}

\item{size}{Size of the drawn sample}
}
\value{
a matrix that is max.time by size
}
\description{
The purpose of this function is to allow a user to simulate abundance
measurements from a fitted GAM model. This is primarily used for testing purposes and not meant
to be called by the user in general.
}
\author{
Devin S. Johnson
}
