% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.simulate.R
\name{site.simulate}
\alias{site.simulate}
\title{Posterior predictive sampling, aggregtion of abundance counts, and linear trend summary}
\usage{
site.simulate(data, obs.formula = NULL, sig.abund = NULL, min.k = 3,
  forecast = 0, family = c("tweedie", "log.normal"), ln.adj = 0,
  obs.prior = NULL, upper = Inf, lower = -Inf, fit.only = FALSE,
  post.size = 1000, run.parallel = FALSE)
}
\arguments{
\item{run.parallel}{}
}
\value{
A named list with the following elements:
\item{site.summary}{A summary of the abundance augmentation for every site and every time.}
\item{posterior.sample}{A named list containing all of the MCMC sample after thinning.}
\item{original.data}{The original data in \code{data}.}
}
\description{
Function for sampling from the posterior predictive distribution of abundance (counts) at
individual sites. Then aggregating the counts over the specified aggregation variable.
}
